feat(usability): handle editing and deleting of existing messages

+++ src/main/resources/db/migration/V8__create_editable_messages_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS editable_message
 (
   chat_id            BIGINT PRIMARY KEY       NOT NULL,
-  message_id         BIGINT                   NOT NULL,
+  message_id         INTEGER                  NOT NULL,
   creation_timestamp TIMESTAMP WITH TIME ZONE NOT NULL
 );
